feat: update the admin user flow
