+++ Gemfile
@@ -54,6 +54,7 @@ group :development, :test do
   gem 'dotenv', '>= 3.0'
   gem 'factory_bot_rails'
   gem 'rspec-rails'
+  gem 'pry'
 end
 
 group :development do
