+++ app/controllers/admin/users_controller.rb
@@ -1,8 +1,10 @@
 module Admin
   class UsersController < BaseController
+    before_action :base_scope
     before_action :set_user, only: %i[show edit update destroy]
 
     def index
+      authorize(User.all)
       @pagy, @users = pagy(User.order(created_at: :desc))
     end
 
@@ -10,6 +12,7 @@ module Admin
 
     def new
       @user = User.new
+      authorize(@user)
     end
 
     def edit; end
@@ -17,8 +20,10 @@ module Admin
     def create
       @user = User.new(user_params)
 
+      authorize(@user)
+
       if @user.save
-        redirect_to admin_users_path(@user), notice: 'User was successfully created.'
+        redirect_to admin_user_path(@user), notice: 'User was successfully created.'
       else
         render :new
       end
@@ -26,7 +31,7 @@ module Admin
 
     def update
       if @user.update(user_params)
-        redirect_to admin_users_path(@user), notice: 'User was successfully updated.'
+        redirect_to admin_user_path(@user), notice: 'User was successfully updated.'
       else
         render :edit
       end
@@ -39,13 +44,17 @@ module Admin
 
     private
 
+    def base_scope
+      policy_scope(User)
+    end
+
     def set_user
       @user = User.find(params[:id])
       authorize(@user)
     end
 
     def user_params
-      params.require(:user).permit(policy(@user).permitted_attributes)
+      params.require(:user).permit(policy([:admin, @user]).permitted_attributes)
     end
   end
 end
