+++ app/views/admin/users/edit.html.slim
@@ -1,0 +1,4 @@
+.container.mx-auto.px-4
+  h2.mt-8.text-2xl.font-bold.leading-9.tracking-tight.text-primary
+    | Edit User
+  = render partial: 'form', locals: { user: @user }
