--- app/views/layouts/_admin_header.html.slim
@@ -1,2 +1,0 @@
-nav class="bg-white border-gray-200 dark:bg-gray-900"
-  | Add Admin Header here...
