+++ app/views/layouts/admin.html.slim
@@ -16,5 +16,5 @@ html[lang="en"]
     = vite_client_tag
     = vite_javascript_tag "admin", "data-turbo-track": "reload"
   body data-controller="icons"
-    = render "layouts/admin_header"
+    = render "layouts/admin/header"
     = yield
